Dynamic-Impact-Tool-Backup/requirements.txt
streamlit
pandas
duckdb
langchain
openai
<<<<<<< HEAD
tiktoken
plotly
seaborn
matplotlib
python-dotenv
langchain-community

=======
scikit-learn
ollama                   
httpx                 
python-dotenv         
loguru                  
langchain-ollama
tabulate
faiss-cpu 
sentence-transformers
fpdf
pdfplumber
langgraph
python-pptx
reportlab
kaleido
duckdb
>>>>>>> e1bab98 (Modified the code)
